requirements/development.txt
tox==4.6.4
twine==6.1.0
build==1.4.3
ipython==8.10.0
ipdb==0.13.9
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
wheel==0.46.2
setuptools==78.1.1
redis==5.0.1
devtools==0.12.2
PyYAML==6.0.2
